+++ lib/src/front_end/ast_node_visitor.dart
@@ -19,6 +19,7 @@ import '../piece/leading_comment.dart';
 import '../piece/list.dart';
 import '../piece/piece.dart';
 import '../piece/type_parameter_bound.dart';
+import '../piece/type_test.dart';
 import '../piece/variable.dart';
 import '../profile.dart';
 import '../source_code.dart';
@@ -184,17 +185,7 @@ final class AstNodeVisitor extends ThrowingAstVisitor<void> with PieceFactory {
 
   @override
   void visitAsExpression(AsExpression node) {
-    writeInfix(
-      node.expression,
-      node.asOperator,
-      node.type,
-      // Don't use Indent.infix after 3.7 because it will flatten the
-      // indentation if the expression occurs in an assignment.
-      // TODO(rnystrom): We probably do want to format `as` the same as other
-      // binary operators, but the tall style already treats them differently,
-      // so leaving that alone for now.
-      indent: style.is3Dot7 ? Indent.infix : Indent.expression,
-    );
+    writeTypeTest(node.expression, node.asOperator, node.type);
   }
 
   @override
@@ -1256,17 +1247,11 @@ final class AstNodeVisitor extends ThrowingAstVisitor<void> with PieceFactory {
 
   @override
   void visitIsExpression(IsExpression node) {
-    writeInfix(
+    writeTypeTest(
       node.expression,
       node.isOperator,
-      operator2: node.notOperator,
+      bang: node.notOperator,
       node.type,
-      // Don't use Indent.infix after 3.7 because it will flatten the
-      // indentation if the expression occurs in an assignment.
-      // TODO(rnystrom): We probably do want to format `as` the same as other
-      // binary operators, but the tall style already treats them differently,
-      // so leaving that alone for now.
-      indent: style.is3Dot7 ? Indent.infix : Indent.expression,
     );
   }